Tidy up CLIP_bagging and mojo code. Move other codes to deprecated
