--- model.py
@@ -1,984 +1,0 @@
-from torchvision import transforms
-import numpy as np
-import torch
-import torch.nn as nn
-import clip
-from dalle2_pytorch import DiffusionPrior
-from dalle2_pytorch.dalle2_pytorch import l2norm, default, exists
-from dalle2_pytorch.dalle2_pytorch import RotaryEmbedding, CausalTransformer, SinusoidalPosEmb, MLP, Rearrange, repeat, rearrange, prob_mask_like, LayerNorm, RelPosBias, Attention, FeedForward
-from dalle2_pytorch.train_configs import DiffusionPriorNetworkConfig
-from tqdm import tqdm
-from functools import partial
-import open_clip
-import os
-import json
-import random
-from diffusers.models.vae import Decoder
-
-class Clipper(torch.nn.Module):
-    def __init__(self, clip_variant='ViT-L/14', hidden_state=False, norm_embs=False, device=torch.device('cpu')):
-        super().__init__()
-        print(clip_variant, device)
-        # clip_model, _, preprocess = open_clip.create_model_and_transforms(clip_variant, pretrained='laion2b_s32b_b82k')
-        clip_model, preprocess = clip.load(clip_variant, device=device)
-        clip_model.eval() # dont want to train model
-        for param in clip_model.parameters():
-            param.requires_grad = False # dont need to calculate gradients
-
-
-        self.clip = clip_model
-        self.preprocess = preprocess
-        self.device = device
-        self.norm_embs = norm_embs
-        self.clamp_embs=False
-        self.clip_size = (425,425)
-        self.preproc = transforms.Compose([
-                transforms.Resize(224, interpolation=transforms.InterpolationMode.BICUBIC, antialias=None),
-                transforms.CenterCrop(224),
-                transforms.Normalize(mean=(0.48145466, 0.4578275, 0.40821073), std=(0.26862954, 0.26130258, 0.27577711))
-        ])
-        if hidden_state:
-            print("THIS IS NOT WORKING CURRENTLY!")
-            clip_model.visual.transformer.resblocks[31].mlp = nn.Identity()
-            clip_model.visual.ln_post = nn.Identity()
-            clip_model.token_embedding = nn.Identity()
-            clip_model.ln_final = nn.Identity()
-        self.mean = np.array([0.48145466, 0.4578275, 0.40821073])
-        self.std = np.array([0.26862954, 0.26130258, 0.27577711])
-        self.normalize = transforms.Normalize(self.mean, self.std)
-    def embed_image(self, image):
-        """Expects images in -1 to 1 range"""
-        clip_emb = self.preprocess(image).unsqueeze(0).to(self.device)
-        clip_emb = self.clip.encode_image(clip_emb)
-        # if self.norm_embs:
-        #     clip_emb = nn.functional.normalize(clip_emb.flatten(1), dim=-1)
-        #     clip_emb = clip_emb.reshape(len(clip_emb),-1,1024)
-        return clip_emb
-    def embed_text(self, text_samples):
-        clip_text = clip.tokenize(text_samples).to(self.device)
-        clip_text = self.clip.encode_text(clip_text)
-        # if self.clamp_embs:
-        #     clip_text = torch.clamp(clip_text, -1.5, 1.5)
-        # if self.norm_embs:
-        #     clip_text = nn.functional.normalize(clip_text, dim=-1)
-        return clip_text
-class OpenClipper(torch.nn.Module):
-    def __init__(self, clip_variant='ViT-L-14', hidden_state=False, norm_embs=False, device=torch.device('cpu')):
-        super().__init__()
-        print(clip_variant, device)
-        # assert clip_variant == 'ViT-L-14' # not setup for other models yet
-         
-        clip_model, _, preprocess = open_clip.create_model_and_transforms(clip_variant, 
-                                        pretrained='laion2b_s32b_b79k', device=device)
-            
-        clip_model.eval() # dont want to train model
-        for param in clip_model.parameters():
-            param.requires_grad = False # dont need to calculate gradients
-            
-        # overwrite preprocess to accept torch inputs instead of PIL Image
-        
-            
-        self.clip = clip_model
-        self.preprocess = preprocess
-        self.preproc = transforms.Compose([
-                transforms.Resize(224, interpolation=transforms.InterpolationMode.BICUBIC, antialias=None),
-                transforms.CenterCrop(224),
-                transforms.Normalize(mean=(0.48145466, 0.4578275, 0.40821073), std=(0.26862954, 0.26130258, 0.27577711))
-        ])
-        self.device = device
-        self.norm_embs = norm_embs
-        
-        if hidden_state:
-            print("THIS IS NOT WORKING CURRENTLY!")
-            clip_model.visual.transformer.resblocks[31].mlp = nn.Identity()
-            clip_model.visual.ln_post = nn.Identity()
-            clip_model.token_embedding = nn.Identity()
-            clip_model.ln_final = nn.Identity()
-            
-    def embed_image(self, image):
-        """Expects images in -1 to 1 range"""
-        clip_emb = self.preprocess(image.to(self.device))
-        clip_emb = self.clip.encode_image(clip_emb)
-        # if self.norm_embs:
-        #     clip_emb = nn.functional.normalize(clip_emb.flatten(1), dim=-1)
-        #     clip_emb = clip_emb.reshape(len(clip_emb),-1,1024)
-        return clip_emb
-class BrainDiffusionPrior(DiffusionPrior):
-    """
-    Differences from original:
-    - Allow for passing of generators to torch random functions
-    - Option to include the voxel2clip model and pass voxels into forward method
-    - Return predictions when computing loss
-    - Load pretrained model from @nousr trained on LAION aesthetics
-    """
-    def __init__(self, *args, **kwargs):
-        voxel2clip = kwargs.pop('voxel2clip', None)
-        super().__init__(*args, **kwargs)
-        self.voxel2clip = voxel2clip
-
-    @torch.no_grad()
-    def p_sample(self, x, t, text_cond = None, self_cond = None, clip_denoised = True, cond_scale = 1.,
-                generator=None):
-        b, *_, device = *x.shape, x.device
-        model_mean, _, model_log_variance, x_start = self.p_mean_variance(x = x, t = t, text_cond = text_cond, self_cond = self_cond, clip_denoised = clip_denoised, cond_scale = cond_scale)
-        if generator is None:
-            noise = torch.randn_like(x)
-        else:
-            #noise = torch.randn_like(x)
-            noise = torch.randn(x.size(), device=x.device, dtype=x.dtype, generator=generator)
-        # no noise when t == 0
-        nonzero_mask = (1 - (t == 0).float()).reshape(b, *((1,) * (len(x.shape) - 1)))
-        pred = model_mean + nonzero_mask * (0.5 * model_log_variance).exp() * noise
-        return pred, x_start
-
-    @torch.no_grad()
-    def p_sample_loop_ddpm(self, shape, text_cond, cond_scale = 1., generator=None):
-        batch, device = shape[0], self.device
-
-        if generator is None:
-            image_embed = torch.randn(shape, device = device)
-        else:
-            image_embed = torch.randn(shape, device = device, generator=generator)
-        x_start = None # for self-conditioning
-
-        if self.init_image_embed_l2norm:
-            image_embed = l2norm(image_embed) * self.image_embed_scale
-
-        for i in tqdm(reversed(range(0, self.noise_scheduler.num_timesteps)), desc='sampling loop time step', total=self.noise_scheduler.num_timesteps, disable=True):
-            times = torch.full((batch,), i, device = device, dtype = torch.long)
-
-            self_cond = x_start if self.net.self_cond else None
-            image_embed, x_start = self.p_sample(image_embed, times, text_cond = text_cond, self_cond = self_cond, cond_scale = cond_scale,
-                                                 generator=generator)
-
-        if self.sampling_final_clamp_l2norm and self.predict_x_start:
-            image_embed = self.l2norm_clamp_embed(image_embed)
-
-        return image_embed
-
-    def p_losses(self, image_embed, times, text_cond, noise = None):
-        noise = default(noise, lambda: torch.randn_like(image_embed))
-
-        image_embed_noisy = self.noise_scheduler.q_sample(x_start = image_embed, t = times, noise = noise)
-
-        self_cond = None
-        if self.net.self_cond and random.random() < 0.5:
-            with torch.no_grad():
-                self_cond = self.net(image_embed_noisy, times, **text_cond).detach()
-
-        pred = self.net(
-            image_embed_noisy,
-            times,
-            self_cond = self_cond,
-            text_cond_drop_prob = self.text_cond_drop_prob,
-            image_cond_drop_prob = self.image_cond_drop_prob,
-            **text_cond
-        )
-
-        if self.predict_x_start and self.training_clamp_l2norm:
-            pred = self.l2norm_clamp_embed(pred)
-
-        if self.predict_v:
-            target = self.noise_scheduler.calculate_v(image_embed, times, noise)
-        elif self.predict_x_start:
-            target = image_embed
-        else:
-            target = noise
-
-        loss = self.noise_scheduler.loss_fn(pred, target)
-        return loss, pred
-
-    def forward(
-        self,
-        text = None,
-        image = None,
-        voxel = None,
-        text_embed = None,      # allow for training on preprocessed CLIP text and image embeddings
-        image_embed = None,
-        text_encodings = None,  # as well as CLIP text encodings
-        *args,
-        **kwargs
-    ):
-        assert exists(text) ^ exists(text_embed) ^ exists(voxel), 'either text, text embedding, or voxel must be supplied'
-        assert exists(image) ^ exists(image_embed), 'either image or image embedding must be supplied'
-        assert not (self.condition_on_text_encodings and (not exists(text_encodings) and not exists(text))), 'text encodings must be present if you specified you wish to condition on it on initialization'
-
-        if exists(voxel):
-            assert exists(self.voxel2clip), 'voxel2clip must be trained if you wish to pass in voxels'
-            assert not exists(text_embed), 'cannot pass in both text and voxels'
-            if self.voxel2clip.use_projector:
-                clip_voxels_mse, clip_voxels = self.voxel2clip(voxel)
-                text_embed = clip_voxels_mse
-            else:
-                clip_voxels = self.voxel2clip(voxel)
-                text_embed = clip_voxels_mse = clip_voxels
-            # text_embed = self.voxel2clip(voxel)
-
-        if exists(image):
-            image_embed, _ = self.clip.embed_image(image)
-
-        # calculate text conditionings, based on what is passed in
-
-        if exists(text):
-            text_embed, text_encodings = self.clip.embed_text(text)
-
-        text_cond = dict(text_embed = text_embed)
-
-        if self.condition_on_text_encodings:
-            assert exists(text_encodings), 'text encodings must be present for diffusion prior if specified'
-            text_cond = {**text_cond, 'text_encodings': text_encodings}
-
-        # timestep conditioning from ddpm
-
-        batch, device = image_embed.shape[0], image_embed.device
-        times = self.noise_scheduler.sample_random_times(batch)
-
-        # PS: I dont think we need this? also if uncommented this does in-place global variable change
-        # scale image embed (Katherine)
-        # image_embed *= self.image_embed_scale
-
-        # calculate forward loss
-
-        loss, pred = self.p_losses(image_embed*self.image_embed_scale, times, text_cond = text_cond, *args, **kwargs)
-
-        # undo the scaling so we can directly use it for real mse loss and reconstruction
-        return loss, pred
-    
-class DiffusionPriorNetwork(nn.Module):
-    def __init__(
-        self,
-        dim,
-        num_timesteps = None,
-        num_time_embeds = 1,
-        # num_image_embeds = 1,
-        # num_brain_embeds = 1,
-        num_tokens = 257,
-        causal = True,
-        learned_query_mode = 'none',
-        **kwargs
-    ):
-        super().__init__()
-        self.dim = dim
-        self.num_time_embeds = num_time_embeds
-        self.continuous_embedded_time = not exists(num_timesteps)
-        self.learned_query_mode = learned_query_mode
-
-        self.to_time_embeds = nn.Sequential(
-            nn.Embedding(num_timesteps, dim * num_time_embeds) if exists(num_timesteps) else nn.Sequential(SinusoidalPosEmb(dim), MLP(dim, dim * num_time_embeds)), # also offer a continuous version of timestep embeddings, with a 2 layer MLP
-            Rearrange('b (n d) -> b n d', n = num_time_embeds)
-        )
-
-        if self.learned_query_mode == 'token':
-            self.learned_query = nn.Parameter(torch.randn(num_tokens, dim))
-        if self.learned_query_mode == 'pos_emb':
-            scale = dim ** -0.5
-            self.learned_query = nn.Parameter(torch.randn(num_tokens, dim) * scale)
-        if self.learned_query_mode == 'all_pos_emb':
-            scale = dim ** -0.5
-            self.learned_query = nn.Parameter(torch.randn(num_tokens*2+1, dim) * scale)
-        self.causal_transformer = FlaggedCausalTransformer(dim = dim, causal=causal, **kwargs)
-
-        self.null_brain_embeds = nn.Parameter(torch.randn(num_tokens, dim))
-        self.null_image_embed = nn.Parameter(torch.randn(num_tokens, dim))
-
-        self.num_tokens = num_tokens
-        self.self_cond = False
-
-    def forward_with_cond_scale(
-        self,
-        *args,
-        cond_scale = 1.,
-        **kwargs
-    ):
-        logits = self.forward(*args, **kwargs)
-
-        if cond_scale == 1:
-            return logits
-
-        null_logits = self.forward(*args, brain_cond_drop_prob = 1., image_cond_drop_prob = 1, **kwargs)
-        return null_logits + (logits - null_logits) * cond_scale
-
-    def forward(
-        self,
-        image_embed,
-        diffusion_timesteps,
-        *,
-        self_cond=None,
-        brain_embed=None,
-        text_embed=None,
-        brain_cond_drop_prob = 0.,
-        text_cond_drop_prob = None,
-        image_cond_drop_prob = 0.
-    ):
-        if text_embed is not None:
-            brain_embed = text_embed
-        if text_cond_drop_prob is not None:
-            brain_cond_drop_prob = text_cond_drop_prob
-
-        image_embed = image_embed.view(len(image_embed),-1,768)
-        # text_embed = text_embed.view(len(text_embed),-1,768)
-        brain_embed = brain_embed.view(len(brain_embed),-1,768)
-        # print(*image_embed.shape)
-        # print(*image_embed.shape, image_embed.device, image_embed.dtype)
-
-        batch, _, dim, device, dtype = *image_embed.shape, image_embed.device, image_embed.dtype
-        # num_time_embeds, num_image_embeds, num_brain_embeds = self.num_time_embeds, self.num_image_embeds, self.num_brain_embeds
-
-        # classifier free guidance masks
-        brain_keep_mask = prob_mask_like((batch,), 1 - brain_cond_drop_prob, device = device)
-        brain_keep_mask = rearrange(brain_keep_mask, 'b -> b 1 1')
-
-        image_keep_mask = prob_mask_like((batch,), 1 - image_cond_drop_prob, device = device)
-        image_keep_mask = rearrange(image_keep_mask, 'b -> b 1 1')
-
-        # mask out brain embeddings with null brain embeddings
-
-        # import pdb; pdb.set_trace()
-        null_brain_embeds = self.null_brain_embeds.to(brain_embed.dtype)
-        brain_embed = torch.where(
-            brain_keep_mask,
-            brain_embed,
-            null_brain_embeds[None]
-        )
-
-        # mask out image embeddings with null image embeddings
-        null_image_embed = self.null_image_embed.to(image_embed.dtype)
-        image_embed = torch.where(
-            image_keep_mask,
-            image_embed,
-            null_image_embed[None]
-        )
-
-        # whether brain embedding is used for conditioning depends on whether brain encodings are available for attention (for classifier free guidance, even though it seems from the paper it was not used in the prior ddpm, as the objective is different)
-        # but let's just do it right
-        if self.continuous_embedded_time:
-            # if continuous cast to flat, else keep int for indexing embeddings
-            diffusion_timesteps = diffusion_timesteps.type(dtype)
-        time_embed = self.to_time_embeds(diffusion_timesteps)
-
-        if self.learned_query_mode == 'token':
-            learned_queries = repeat(self.learned_query, 'n d -> b n d', b = batch)
-        elif self.learned_query_mode == 'pos_emb':
-            pos_embs = repeat(self.learned_query, 'n d -> b n d', b = batch)
-            image_embed = image_embed + pos_embs
-            learned_queries = torch.empty((batch, 0, dim), device=brain_embed.device)
-        elif self.learned_query_mode == 'all_pos_emb':
-            pos_embs = repeat(self.learned_query, 'n d -> b n d', b = batch)
-            learned_queries = torch.empty((batch, 0, dim), device=brain_embed.device)
-        else:
-            learned_queries = torch.empty((batch, 0, dim), device=brain_embed.device)
-
-        tokens = torch.cat((
-            brain_embed,  # 257
-            time_embed,  # 1
-            image_embed,  # 257
-            learned_queries  # 257
-        ), dim = -2)
-        if self.learned_query_mode == 'all_pos_emb':
-            tokens = tokens + pos_embs
-
-        # attend
-        tokens = self.causal_transformer(tokens)
-
-        # get learned query, which should predict the image embedding (per DDPM timestep)
-        pred_image_embed = tokens[..., -self.num_tokens:, :]
-
-        return pred_image_embed
-class BrainNetwork(nn.Module):
-    def __init__(self, out_dim=768, in_dim=39548, clip_size=768, h=2048, n_blocks=4, norm_type='ln', act_first=False, use_projector=True):
-        super().__init__()
-        norm_func = partial(nn.BatchNorm1d, num_features=h) if norm_type == 'bn' else partial(nn.LayerNorm, normalized_shape=h)
-        act_fn = partial(nn.ReLU, inplace=True) if norm_type == 'bn' else nn.GELU
-        act_and_norm = (act_fn, norm_func) if act_first else (norm_func, act_fn)
-        # self.temp = nn.Parameter(torch.tensor(.006))
-        self.lin0 = nn.Sequential(
-            nn.Linear(in_dim, h),
-            *[item() for item in act_and_norm],
-            nn.Dropout(0.5),
-        )
-        self.mlp = nn.ModuleList([
-            nn.Sequential(
-                nn.Linear(h, h),
-                *[item() for item in act_and_norm],
-                nn.Dropout(0.15)
-            ) for _ in range(n_blocks)
-        ])
-        self.lin1 = nn.Linear(h, out_dim, bias=True)
-        self.n_blocks = n_blocks
-        self.clip_size = clip_size
-
-        self.use_projector = use_projector
-        if use_projector:
-            self.projector = nn.Sequential(
-                nn.LayerNorm(clip_size),
-                nn.GELU(),
-                nn.Linear(clip_size, 2048),
-                nn.LayerNorm(2048),
-                nn.GELU(),
-                nn.Linear(2048, 2048),
-                nn.LayerNorm(2048),
-                nn.GELU(),
-                nn.Linear(2048, clip_size)
-            )
-
-    def forward(self, x):
-        '''
-            bs, 1, 15724 -> bs, 32, h
-            bs, 32, h -> bs, 32h
-            b2, 32h -> bs, 768
-        '''
-        if x.ndim == 4:
-            # case when we passed 3D data of shape [N, 81, 104, 83]
-            assert x.shape[1] == 81 and x.shape[2] == 104 and x.shape[3] == 83
-            # [N, 699192]
-            x = x.reshape(x.shape[0], -1)
-
-        x = self.lin0(x)  # bs, h
-        residual = x
-        for res_block in range(self.n_blocks):
-            x = self.mlp[res_block](x)
-            x = x + residual
-            residual = x
-        x = x.reshape(len(x), -1)
-        x = self.lin1(x)
-        if self.use_projector:
-            return x, self.projector(x.reshape(len(x), -1, self.clip_size))
-        return x
-class FlaggedCausalTransformer(nn.Module):
-    def __init__(
-        self,
-        *,
-        dim,
-        depth,
-        dim_head = 64,
-        heads = 8,
-        ff_mult = 4,
-        norm_in = False,
-        norm_out = True,
-        attn_dropout = 0.,
-        ff_dropout = 0.,
-        final_proj = True,
-        normformer = False,
-        rotary_emb = True,
-        causal=True
-    ):
-        super().__init__()
-        self.init_norm = LayerNorm(dim) if norm_in else nn.Identity() # from latest BLOOM model and Yandex's YaLM
-
-        self.rel_pos_bias = RelPosBias(heads = heads)
-
-        rotary_emb = RotaryEmbedding(dim = min(32, dim_head)) if rotary_emb else None
-
-        self.layers = nn.ModuleList([])
-        for _ in range(depth):
-            self.layers.append(nn.ModuleList([
-                Attention(dim = dim, causal = causal, dim_head = dim_head, heads = heads, dropout = attn_dropout, rotary_emb = rotary_emb),
-                FeedForward(dim = dim, mult = ff_mult, dropout = ff_dropout, post_activation_norm = normformer)
-            ]))
-
-        self.norm = LayerNorm(dim, stable = True) if norm_out else nn.Identity()  # unclear in paper whether they projected after the classic layer norm for the final denoised image embedding, or just had the transformer output it directly: plan on offering both options
-        self.project_out = nn.Linear(dim, dim, bias = False) if final_proj else nn.Identity()
-
-    def forward(self, x):
-        n, device = x.shape[1], x.device
-
-        x = self.init_norm(x)
-
-        attn_bias = self.rel_pos_bias(n, n + 1, device = device)
-
-        for attn, ff in self.layers:
-            x = attn(x, attn_bias = attn_bias) + x
-            x = ff(x) + x
-
-        out = self.norm(x)
-        return self.project_out(out)
-class BrainDiffusionPriorOld(DiffusionPrior):
-    """ 
-    Differences from original:
-    - Allow for passing of generators to torch random functions
-    - Option to include the voxel2clip model and pass voxels into forward method
-    - Return predictions when computing loss
-    - Load pretrained model from @nousr trained on LAION aesthetics
-    """
-    def __init__(self, *args, **kwargs):
-        voxel2clip = kwargs.pop('voxel2clip', None)
-        super().__init__(*args, **kwargs)
-        self.voxel2clip = voxel2clip
-
-    @torch.no_grad()
-    def p_sample(self, x, t, text_cond = None, self_cond = None, clip_denoised = True, cond_scale = 1.,
-                generator=None):
-        b, *_, device = *x.shape, x.device
-        model_mean, _, model_log_variance, x_start = self.p_mean_variance(x = x, t = t, text_cond = text_cond, self_cond = self_cond, clip_denoised = clip_denoised, cond_scale = cond_scale)
-        if generator is None:
-            noise = torch.randn_like(x)
-        else:
-            #noise = torch.randn_like(x)
-            noise = torch.randn(x.size(), device=x.device, dtype=x.dtype, generator=generator)
-        # no noise when t == 0
-        nonzero_mask = (1 - (t == 0).float()).reshape(b, *((1,) * (len(x.shape) - 1)))
-        pred = model_mean + nonzero_mask * (0.5 * model_log_variance).exp() * noise
-        return pred, x_start
-
-    @torch.no_grad()
-    def p_sample_loop_ddpm(self, shape, text_cond, cond_scale = 1., generator=None):
-        batch, device = shape[0], self.device
-
-        if generator is None:
-            image_embed = torch.randn(shape, device = device)
-        else:
-            image_embed = torch.randn(shape, device = device, generator=generator)
-        x_start = None # for self-conditioning
-
-        if self.init_image_embed_l2norm:
-            image_embed = l2norm(image_embed) * self.image_embed_scale
-
-        for i in tqdm(reversed(range(0, self.noise_scheduler.num_timesteps)), desc='sampling loop time step', total=self.noise_scheduler.num_timesteps, disable=True):
-            times = torch.full((batch,), i, device = device, dtype = torch.long)
-
-            self_cond = x_start if self.net.self_cond else None
-            image_embed, x_start = self.p_sample(image_embed, times, text_cond = text_cond, self_cond = self_cond, cond_scale = cond_scale, 
-                                                 generator=generator)
-
-        if self.sampling_final_clamp_l2norm and self.predict_x_start:
-            image_embed = self.l2norm_clamp_embed(image_embed)
-
-        return image_embed
-
-    def p_losses(self, image_embed, times, text_cond, noise = None):
-        noise = default(noise, lambda: torch.randn_like(image_embed))
-
-        image_embed_noisy = self.noise_scheduler.q_sample(x_start = image_embed, t = times, noise = noise)
-
-        self_cond = None
-        if self.net.self_cond and random.random() < 0.5:
-            with torch.no_grad():
-                self_cond = self.net(image_embed_noisy, times, **text_cond).detach()
-
-        pred = self.net(
-            image_embed_noisy,
-            times,
-            self_cond = self_cond,
-            text_cond_drop_prob = self.text_cond_drop_prob,
-            image_cond_drop_prob = self.image_cond_drop_prob,
-            **text_cond
-        )
-
-        if self.predict_x_start and self.training_clamp_l2norm:
-            pred = self.l2norm_clamp_embed(pred)
-
-        if self.predict_v:
-            target = self.noise_scheduler.calculate_v(image_embed, times, noise)
-        elif self.predict_x_start:
-            target = image_embed
-        else:
-            target = noise
-
-        loss = self.noise_scheduler.loss_fn(pred, target)
-        return loss, pred
-
-    def forward(
-        self,
-        text = None,
-        image = None,
-        voxel = None,
-        text_embed = None,      # allow for training on preprocessed CLIP text and image embeddings
-        image_embed = None,
-        text_encodings = None,  # as well as CLIP text encodings
-        *args,
-        **kwargs
-    ):
-        assert exists(text) ^ exists(text_embed) ^ exists(voxel), 'either text, text embedding, or voxel must be supplied'
-        assert exists(image) ^ exists(image_embed), 'either image or image embedding must be supplied'
-        assert not (self.condition_on_text_encodings and (not exists(text_encodings) and not exists(text))), 'text encodings must be present if you specified you wish to condition on it on initialization'
-
-        if exists(voxel):
-            assert exists(self.voxel2clip), 'voxel2clip must be trained if you wish to pass in voxels'
-            assert not exists(text_embed), 'cannot pass in both text and voxels'
-            text_embed = self.voxel2clip(voxel)
-
-        if exists(image):
-            image_embed, _ = self.clip.embed_image(image)
-
-        # calculate text conditionings, based on what is passed in
-
-        if exists(text):
-            text_embed, text_encodings = self.clip.embed_text(text)
-
-        text_cond = dict(text_embed = text_embed)
-
-        if self.condition_on_text_encodings:
-            assert exists(text_encodings), 'text encodings must be present for diffusion prior if specified'
-            text_cond = {**text_cond, 'text_encodings': text_encodings}
-
-        # timestep conditioning from ddpm
-
-        batch, device = image_embed.shape[0], image_embed.device
-        times = self.noise_scheduler.sample_random_times(batch)
-
-        # scale image embed (Katherine)
-
-        image_embed *= self.image_embed_scale
-
-        # calculate forward loss
-
-        loss, pred = self.p_losses(image_embed, times, text_cond = text_cond, *args, **kwargs)
-
-        return loss, pred#, text_embed
-   
-    @staticmethod
-    def from_pretrained(net_kwargs={}, prior_kwargs={}, voxel2clip_path=None, ckpt_dir='./checkpoints'):
-        # "https://huggingface.co/nousr/conditioned-prior/raw/main/vit-l-14/aesthetic/prior_config.json"
-        config_url = os.path.join(ckpt_dir, "prior_config.json")
-        config = json.load(open(config_url))
-        
-        config['prior']['net']['max_text_len'] = 256
-        config['prior']['net'].update(net_kwargs)
-        # print('net_config', config['prior']['net'])
-        net_config = DiffusionPriorNetworkConfig(**config['prior']['net'])
-
-        kwargs = config['prior']
-        kwargs.pop('clip')
-        kwargs.pop('net')
-        kwargs.update(prior_kwargs)
-        # print('prior_config', kwargs)
-
-        diffusion_prior_network = net_config.create()
-        diffusion_prior = BrainDiffusionPriorOld(net=diffusion_prior_network, clip=None, **kwargs).to(torch.device('cpu'))
-        
-        # 'https://huggingface.co/nousr/conditioned-prior/resolve/main/vit-l-14/aesthetic/best.pth'
-        ckpt_url = os.path.join(ckpt_dir, 'best.pth')
-        ckpt = torch.load(ckpt_url, map_location=torch.device('cpu'))
-
-        # Note these keys will be missing (maybe due to an update to the code since training):
-        # "net.null_text_encodings", "net.null_text_embeds", "net.null_image_embed"
-        # I don't think these get used if `cond_drop_prob = 0` though (which is the default here)
-        diffusion_prior.load_state_dict(ckpt, strict=False)
-        # keys = diffusion_prior.load_state_dict(ckpt, strict=False)
-        # print("missing keys in prior checkpoint (probably ok)", keys.missing_keys)
-
-        if voxel2clip_path:
-            # load the voxel2clip weights
-            checkpoint = torch.load(voxel2clip_path, map_location=torch.device('cpu'))
-            
-            state_dict = checkpoint['model_state_dict']
-            for key in list(state_dict.keys()):
-                if 'module.' in key:
-                    state_dict[key.replace('module.', '')] = state_dict[key]
-                    del state_dict[key]
-            diffusion_prior.voxel2clip.load_state_dict(state_dict)
-        
-        return diffusion_prior
-    
-class Voxel2StableDiffusionModel(torch.nn.Module):
-    # define the prototype of the module
-    def __init__(self, in_dim=39548, h=2048, n_blocks=4):
-        super().__init__()
-
-        self.lin0 = nn.Sequential(
-            nn.Linear(in_dim, h, bias=False),
-            nn.LayerNorm(h),
-            nn.SiLU(inplace=True),
-            nn.Dropout(0.5),
-        )
-
-        self.mlp = nn.ModuleList([
-            nn.Sequential(
-                nn.Linear(h, h, bias=False),
-                nn.LayerNorm(h),
-                nn.SiLU(inplace=True),
-                nn.Dropout(0.3),
-            ) for _ in range(n_blocks)
-        ])
-        
-        self.lin1 = nn.Linear(h, 16384, bias=False)
-        self.norm = nn.GroupNorm(1, 64)
-
-        def init_weights(m):
-            if isinstance(m, nn.Linear):
-                torch.nn.init.xavier_uniform_(m.weight)
-
-        init_weights(self.lin0)
-        self.mlp.apply(init_weights)
-        init_weights(self.lin1)
-
-        self.upsampler = Decoder(
-            in_channels=64,
-            out_channels=4,
-            up_block_types=["UpDecoderBlock2D","UpDecoderBlock2D","UpDecoderBlock2D"],
-            block_out_channels=[64, 128, 256],
-            layers_per_block=1,
-        )
-        for parm in self.upsampler.parameters():
-            parm.require_grad = False
-        self.upsampler.eval()
-
-    # define how it forward, using the module defined above
-    def forward(self, x):
-        x = self.lin0(x)
-        residual = x
-        for res_block in self.mlp:
-            x = res_block(x)
-            x = x + residual
-            residual = x
-        x = x.reshape(len(x), -1)
-        x = self.lin1(x)
-        x = self.norm(x.reshape(x.shape[0], -1, 16, 16).contiguous())
-        return self.upsampler(x)
-
-from transformers import BlipImageProcessor
-
-class Voxel2Blip(torch.nn.Module):
-    # define the prototype of the module
-    def __init__(self, in_dim=39548, h=4096, n_blocks=4):
-        super().__init__()
-
-        self.lin0 = nn.Sequential(
-            nn.Linear(in_dim, h, bias=False),
-            nn.LayerNorm(h),
-            nn.ReLU(inplace=True),
-            nn.Dropout(0.3),
-        )
-        self.mlp = nn.ModuleList([
-            nn.Sequential(
-                nn.Linear(h, h, bias=False),
-                nn.LayerNorm(h),
-                nn.ReLU(inplace=True),
-                nn.Dropout(0.3),
-            ) for _ in range(n_blocks)
-        ])
-        self.lin1 = nn.Linear(h, 16384, bias=False)
-        self.norm = nn.GroupNorm(1, 64)
-
-        def init_weights(m):
-            if isinstance(m, nn.Linear):
-                torch.nn.init.xavier_uniform_(m.weight)
-
-        init_weights(self.lin0)
-        self.mlp.apply(init_weights)
-        init_weights(self.lin1)
-
-        self.upsampler = Decoder(
-            in_channels=64,
-            out_channels=4,
-            up_block_types=["UpDecoderBlock2D","UpDecoderBlock2D","UpDecoderBlock2D"],
-            block_out_channels=[64, 128, 256],
-            layers_per_block=1,
-        )
-        for parm in self.upsampler.parameters():
-            parm.require_grad = False
-        self.upsampler.eval()
-
-    # define how it forward, using the module defined above
-    def forward(self, x):
-        x = self.lin0(x)
-        residual = x
-        for res_block in self.mlp:
-            x = res_block(x)
-            x = x + residual
-            residual = x
-        x = x.reshape(len(x), -1)
-        x = self.lin1(x)
-        x = self.norm(x.reshape(x.shape[0], -1, 16, 16).contiguous())
-        return self.upsampler(x)
-
-    
-# class Voxel2Blip(torch.nn.Module):
-#     # define the prototype of the module
-#     def __init__(self, in_dim=39548, h=2048, n_blocks=4):
-#         super().__init__()
-
-#         self.lin0 = nn.Sequential(
-#             nn.Linear(in_dim, h),
-#             nn.LayerNorm(h),
-#             nn.ReLU(inplace=True),
-#             nn.Dropout(0.3),
-#         )
-        
-        
-#         self.mlp = nn.ModuleList([
-#             nn.Sequential(
-#                 nn.Linear(h, h),
-#                 nn.LayerNorm(h),
-#                 nn.ReLU(inplace=True),
-#                 nn.Dropout(0.3),
-#             ) for _ in range(n_blocks)
-#         ])
-#         self.lin1 = nn.Sequential(
-#                 nn.Linear(h, 9216, bias=False),
-#                 nn.ReLU(inplace=True),
-#             ) 
-        
-#         self.norm = nn.GroupNorm(1, 4)
-
-#     # define how it forward, using the module defined above
-#     def forward(self, x):
-#         x = self.lin0(x)
-#         residual = x
-#         for res_block in self.mlp:
-#             x = res_block(x)
-#             x = x + residual
-#             residual = x
-#         x = x.reshape(len(x), -1)
-#         x = self.lin1(x)
-#         x = self.norm(x.reshape(x.shape[0],4,48,48).contiguous())
-#         # x = x.reshape(x.shape[0],4,48,48)
-#         return x
-
-# class Voxel2Blip(torch.nn.Module):
-#     # define the prototype of the module
-#     def __init__(self, in_dim=39548, h=2048, n_blocks=2):
-#         super().__init__()
-
-#         self.lin0 = nn.Sequential(
-#             nn.Linear(in_dim, 4*h),
-#             # nn.LayerNorm(4*h),
-#             nn.ReLU(inplace=True),
-#             nn.Dropout(0.2),
-#         )
-        
-#         self.lin1 = nn.Sequential(
-#                 nn.Linear(4*h, 2*h),
-#                 # nn.LayerNorm(2*h),
-#                 nn.ReLU(inplace=True),
-#                 nn.Dropout(0.2),
-#             )
-#         self.lin2 = nn.Sequential(
-#                 nn.Linear(2*h, h),
-#                 # nn.LayerNorm(h),
-#                 nn.ReLU(inplace=True),
-#                 nn.Dropout(0.2),
-#             )
-#         self.mlp = nn.ModuleList([
-#             nn.Sequential(
-#                 nn.Linear(h, h),
-#                 # nn.LayerNorm(h),
-#                 nn.ReLU(inplace=True),
-#                 nn.Dropout(0.2),
-#             ) for _ in range(n_blocks)
-#         ])
-#         self.lin_out = nn.Sequential(
-#                 nn.Linear(h, 9216, bias=False),
-#                 nn.ReLU(inplace=True),
-#             ) 
-        
-#         self.norm = nn.GroupNorm(1, 4)
-
-#     # define how it forward, using the module defined above
-#     def forward(self, x):
-#         x = self.lin0(x)
-#         x = self.lin1(x)
-#         x = self.lin2(x)
-#         residual = x
-#         for res_block in self.mlp:
-#             x = res_block(x)
-#             x = x + residual
-#             residual = x
-#         x = x.reshape(len(x), -1)
-#         x = self.lin_out(x)
-#         x = self.norm(x.reshape(x.shape[0],4,48,48).contiguous())
-#         # x = x.reshape(x.shape[0],4,48,48)
-#         return x
-
-class Voxel2Blip(torch.nn.Module):
-    # define the prototype of the module
-    def __init__(self, in_dim=39548, h=4096, n_blocks=3):
-        super().__init__()
-
-        self.lin0 = nn.Sequential(
-            nn.Linear(in_dim, h),
-            # nn.LayerNorm(4*h),
-            nn.SiLU(inplace=True),
-            nn.Dropout(0.2),
-        )
-        # self.lin2 = nn.Sequential(
-        #         nn.Linear(2*h, h),
-        #         # nn.LayerNorm(h),
-        #         nn.ReLU(inplace=True),
-        #         nn.Dropout(0.2),
-        #     )
-        self.mlp = nn.ModuleList([
-            nn.Sequential(
-                nn.Linear(h, h),
-                # nn.LayerNorm(h),
-                nn.SiLU(inplace=True),
-                nn.Dropout(0.2),
-            ) for _ in range(n_blocks)
-        ])
-        self.lin_out = nn.Sequential(
-                nn.Linear(h, 9216, bias=False),
-                nn.SiLU(inplace=True),
-            ) 
-        
-        self.norm = nn.GroupNorm(1, 4)
-
-    # define how it forward, using the module defined above
-    def forward(self, x):
-        x = self.lin0(x)
-        # x = self.lin1(x)
-        # x = self.lin2(x)
-        residual = x
-        for res_block in self.mlp:
-            x = res_block(x)
-            x = x + residual
-            residual = x
-        x = x.reshape(len(x), -1)
-        x = self.lin_out(x)
-        x = self.norm(x.reshape(x.shape[0],4,48,48).contiguous())
-        # x = x.reshape(x.shape[0],4,48,48)
-        return x
-    
-class Voxel2Classification(torch.nn.Module):
-    # define the prototype of the module
-    def __init__(self, in_dim=39548,out_dim = 133, h=4096, n_blocks=2):
-        super().__init__()
-
-        self.lin0 = nn.Sequential(
-            nn.Linear(in_dim, h),
-            # nn.LayerNorm(4*h),
-            nn.ReLU(inplace=True),
-            nn.Dropout(0.2),
-        )
-        self.lin1 = nn.Sequential(
-                nn.Linear(h, 2048),
-                # nn.LayerNorm(h),
-                nn.ReLU(inplace=True),
-                nn.Dropout(0.2),
-            )
-        self.lin2 = nn.Sequential(
-                nn.Linear(2048, 512),
-                # nn.LayerNorm(h),
-                nn.ReLU(inplace=True),
-                nn.Dropout(0.2),
-            )
-        self.mlp = nn.ModuleList([
-            nn.Sequential(
-                nn.Linear(2048, 2048),
-                # nn.LayerNorm(h),
-                nn.ReLU(inplace=True),
-                nn.Dropout(0.2),
-            ) for _ in range(n_blocks)
-        ])
-        self.lin_out = nn.Sequential(
-                nn.Linear(512, out_dim, bias=False),
-                nn.Sigmoid(),
-            ) 
-        
-        # self.norm = nn.GroupNorm(1, 4)
-
-    # define how it forward, using the module defined above
-    def forward(self, x):
-        x = self.lin0(x)
-        x = self.lin1(x)
-        
-        residual = x
-        for res_block in self.mlp:
-            x = res_block(x)
-            x = x + residual
-            residual = x
-        x = x.reshape(len(x), -1)
-        x = self.lin2(x)
-        x = self.lin_out(x)
-        # x = self.norm(x.reshape(x.shape[0],4,48,48).contiguous())
-        # x = x.reshape(x.shape[0],4,48,48)
-        return x
