+++ requirements.txt
@@ -1,0 +1,8 @@
+numpy==1.22.2
+torch==2.1.2
+torchsummary==1.5.1
+torchvision==0.16.2
+diffusers==0.24.0
+wandb==0.16.1
+Pillow==10.1.0
+transformers==4.36.2
